Update site connection IKE/ESP config options
This adds removes the required flag from all
currently required IKE/ESP config options, in
favor of defaulting the values to the current
server defined defaults.

In some cases special logic is required to identify
values that are not applicable in the IKE V1 case,
or whether an integrity algorithm should be supplied.