adds pureport.aws collection with patched modules
This commit adds the pureprot.aws collection that contains two modules
that are required to fully automate AWS DirectConnect connections.  Both
modules have been sumitted upstream to the Ansible AWS community and are
pending review and acceptance.  This collection serves as an
intermediate solution and will be removed once the upstream work is
merged.

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>