adds support for gcp subnetworks (#25)
* adds support for gcp subnetworks

This commit adds support for configuring gcp subnetworks to the
google_cloud_interconnect role.  Playbook authors can now choose to have
subnets automatically created for the vpc or to create a custom subnet
and attach it to the vpc.  Playbook authors can also now choose between
GLOBAL or REGIONAL routing modes for the vpc.

This commit also changes the default values to make them a litle more
descriptive.

* fix missed diff statement from rebase