new role `google_cloud_vpc` (#31)
* new role `google_cloud_vpc`

This commit adds a new role to the collection for creating google cloud
vpc networks.  This role allows the vpc to be managed as well as any
firewall rules.

* fix lint exposed errors

* removes configurable defaults

This commit changes the role implementation to autogenerate the values
to create google cloud networks.  This role can now run idempontently
with no values being configured.

* fix for gcp_subnets

This fixes a bug with iterating over gcp subnets and removing them