updates aws_direct_connect role to min viable (#32)
* updates aws_direct_connect role to min viable

This commit changes the `aws_direct_connect` role to only implement the
minimium viable resources necessary to enable direct connect.  This
removes the role from creating a vpc and vgw

* update aws_direct_connect role

This update to the role will now generate all necessary values based on
the name of the fabric to connect to. If the value of fabric is not
provided then `default` is used.